biggif/requirements.txt
numpy
matplotlib
h5py
pillow
ipython
opencv-python
